[DEV-1.2] Implemented conirmation view when deletion is requested

+++ assignment2/views/DeleteView.java
@@ -6,10 +6,7 @@ import javafx.collections.ObservableList;
 import javafx.geometry.Insets;
 import javafx.geometry.Pos;
 import javafx.scene.Scene;
-import javafx.scene.control.Button;
-import javafx.scene.control.Label;
-import javafx.scene.control.ListView;
-import javafx.scene.control.SelectionMode;
+import javafx.scene.control.*;
 import javafx.scene.layout.VBox;
 import javafx.scene.text.Font;
 import javafx.stage.Modality;
@@ -145,8 +142,17 @@ public class DeleteView {
      */
     public void deleteGame(String GameFile) throws IOException, ClassNotFoundException {
         File file = new File(GameFile);
-        file.delete();
-        selectGameLabel.setText("Successfully deleted the game!");
+
+        Alert alert = new Alert(Alert.AlertType.CONFIRMATION, "Delete " + GameFile + " ?", ButtonType.YES,  ButtonType.CANCEL);
+        alert.showAndWait();
+
+        if (alert.getResult() == ButtonType.YES) {
+            file.delete();
+            selectGameLabel.setText("Successfully deleted the game!");
+        } else {
+            selectGameLabel.setText("Deletion cancelled!");
+        }
+
         this.adventureGameView.updateScene("");
         getFiles(GameList); //get files for file selector
     }
